requirements.txt
# T2DM Persistence RWE — Python 3.11 dependencies
# Core data
duckdb==0.10.3
pandas==2.2.2
numpy==1.26.4
openpyxl==3.1.2

# Survival analysis
lifelines==0.27.8

# ML
xgboost==2.0.3
scikit-learn==1.5.0
shap==0.45.1
umap-learn==0.5.6

# Visualization
matplotlib==3.9.0
seaborn==0.13.2
plotly==5.22.0
kaleido==0.2.1

# Graph
networkx==3.3
neo4j==5.20.0

# LLM / chatbot
anthropic==0.28.0
langchain==0.2.5
langchain-anthropic==0.1.15
langchain-community==0.2.5
faiss-cpu==1.8.0

# Dashboard
streamlit==1.35.0

# Jupyter
jupyter==1.1.1
notebook==7.2.1
ipykernel==6.29.5
nbformat==5.10.4
nbconvert==7.17.1

# Utilities
scipy==1.13.1
statsmodels==0.14.2
python-dotenv==1.0.1
tqdm==4.66.4
PyYAML==6.0.1
requests==2.32.3
Pillow>=10.0.0
